+++ frontend/components/reviews/review.jsx
@@ -24,8 +24,9 @@ class Review extends React.Component {
     }
 
     handleDelete(e) {
-        debugger
-        this.props.deleteReview(e.target.value);
+        if (this.props.currUserId === e.target.getAttribute('data-value')) {
+            this.props.deleteReview(e.target.value);
+        }
     }
 
     displayConditions(review) {
@@ -108,7 +109,7 @@ class Review extends React.Component {
                                 <span className='rev-description'>{rev.description}</span>
                             </section>
                             <section>
-                                <button onClick={this.handleDelete} value={rev.id}>Delete</button>
+                                <button onClick={this.handleDelete} value={rev.id} data-value={rev.user_id}>Delete</button>
                             </section>
                         </div>
                     )
